requirements.txt
setuptools>=45.2.0
python-olm>=3.1.3
matrix-nio[e2e]>=0.8.0
Markdown>=3.1.1
PyYAML>=5.1.2
httpx>=0.11.1
pytz==2019.3
adguardhome>=0.4.1
